pre mock slurm
