+++ main/local_mock/config/nxf_local.nf
@@ -1,0 +1,8 @@
+// parameter defaults
+params.output = 'results'
+
+env {
+    NUMBA_CACHE_DIR = './temp/numba_cache'
+    MPLCONFIGDIR = './temp/matplotlib'
+    XDG_CACHE_HOME = './temp/xdg_home'
+}
