+++ main/nextflow.composition/test.sh
@@ -8,8 +8,9 @@ cd ./cache
 # rm -rf .nextflow* work results
 rm -r results
     # --bootstrap 
+    # run ../test.2.nf \
 nextflow -C ../config.nf -log ./.nextflow_logs/log \
-    run ../test.2.nf \
+    run ../mock.nf \
     --given_OVtA given/2beaver_fosmid_seqs.fna given/contigs.fna \
     --given_bAYL given/swissprot_fastal_ref \
-    --account ***REMOVED***
+    --account asdf
