+++ src/metasmith/__init__.py
@@ -1,2 +1,2 @@
 # from .transforms import RegisterTransform
-from .models import ExecutionContext, ExecutionResult, DataInstance, DataTypeLibrary, TransformInstance
+from .models.libraries import ExecutionContext, ExecutionResult, DataInstance, DataTypeLibrary, TransformInstance
