+++ src/metasmith/transforms.py
@@ -3,7 +3,7 @@ import yaml
 from pathlib import Path
 from typing import Callable, Iterable
 from importlib import import_module, __import__
-from .models import ExecutionContext, ExecutionResult, TransformInstance
+from .models.libraries import ExecutionContext, ExecutionResult, TransformInstance
 
 _last_transform: TransformInstance = None
 def LoadTransform(definition: Path):
